+++ contracts/vending-minter/src/testing.rs
@@ -1,0 +1,2 @@
+mod setup;
+mod tests;
