+++ contracts/vending-minter/src/testing/setup.rs
@@ -1,0 +1,3 @@
+pub mod helpers;
+pub mod setup_accounts_and_block;
+pub mod setup_contracts;
