+++ contracts/vending-minter/src/testing/setup/helpers.rs
@@ -1,0 +1,8 @@
+use cw4::Member;
+
+pub fn member<T: Into<String>>(addr: T, weight: u64) -> Member {
+    Member {
+        addr: addr.into(),
+        weight,
+    }
+}
