+++ contracts/vending-minter/src/testing/tests.rs
@@ -1,0 +1,2 @@
+mod integration_tests;
+mod test_splits;
